テスト分割、テスト用のディレクトリ階層を整理
