--- 05.wc/test_directory/hoge1.txt
@@ -1,1 +1,0 @@
-hoge
