--- 05.wc/test_directory/hoge2.txt
@@ -1,1 +1,0 @@
-hogehoge
