+++ 05.wc/test_directory/test_wc.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'minitest/autorun'
+require_relative '../wc'
+require_relative 'tests/test_parse_options'
+require_relative 'tests/test_count_content'
+require_relative 'tests/test_output_format'
+require_relative 'tests/test_read_and_count_file'
+require_relative 'tests/test_collect_file_stats'
+require_relative 'tests/test_main_pipe'
+require_relative 'tests/test_main_files'
